Safe handling of invalid inputs in delay_estimator.
The delay_estimator crash on invalid create inputs when running new unit tests. This can't occur on a higher level unless corresponding enum and defines are incorrectly changed.

The create and free functions are now more like malloc() and free() in design. The complete change to that will be done in a seperate CL.

BUG=
TEST=

Review URL: https://webrtc-codereview.appspot.com/492003