Disable WEBRTC_SVNREVISION
Description:
This CL will temporally define WEBRTC_SVNREVISION as "n/a" because it
could break Chromium if svn is not installed.
The long term solution is a have a script that could deal with it, and
have it support git-svn which is used by most developers.

BUG=496
TEST=build on Linux
Review URL: https://webrtc-codereview.appspot.com/569007