Continuing to rewrite custom calls.
Added new helper for getting regular input. Rewrote remaining input handling for custom calls. Added a --choose_defaults flag which makes it possible to default on everything (e.g. with the flag, choosing custom call will accept all defaults and go directly to the call).

The next patch will add support for overriding arbitrary choices using flags. That is the point I want to arrive at and this patch paves the way for that. Fortunately it gets rid of some repetitive and bug-prone code on the way.

BUG=

Review URL: https://webrtc-codereview.appspot.com/858005