Update precompiled Android tools for 64-bit platforms (take 2)
In https://review.webrtc.org/15649004 fischman@ provided
good comments on improvements to the instructions for how
to build and update these tools. He also discovered a typo that
led to that the wrong binaries were deployed.

This CL addresses the fixes in the instructions and updates the binaries
again.

R=fischman@chromium.org

Review URL: https://webrtc-codereview.appspot.com/15649006