GN: Fix build by disabling compiler warning in base.
It seems like it is not possible to disabled the -Wall
warnings that are enabled in build/config/compiler/BUILD.gn
with -Wno-all.

According to the documentation at
https://code.google.com/p/chromium/wiki/GNCookbook
the proper way is to disable the chromium_code config instead.

System wrappers also needed some minor fixes for Android.

TBR=henrike@webrtc.org
BUG=3441
TEST=Passing our GN trybots.

Review URL: https://webrtc-codereview.appspot.com/18649004