Remove WebRtcTextureVideoFrame
WebRtcTextureVideoFrame is currently an empty shell that only provides a convenience constructor of I420VideoFrame with a texture buffer. This CL moves that constructor, and all unittests, of WebRtcTextureVideoFrame into the base class. Then it's possible to completely remove WebRtcTextureVideoFrame and all its files.

BUG=1128
R=pbos@webrtc.org, perkj@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/48399004